Altered the subtraction rountines to display their derived stats before and after subtraction, i.e. bias mean: 2333 bias mean after bias subtraction: 0
added an option to get_ds9_region() to specify the bias section. This
should of been added when bias capabilities were added.

Changed background subtract so that it will optionally display the mask
it used to cover any sources.

Added behavior to bias_from_ds9 so that it automatcially aquires a ds9
target if none is given.

Made various changes to variable names, code comments, and doc strings
to improve readability.