Wrote a method for writing the average frames onto disk as a fits file.
It might be worth wrapping the method as it's own function, but the
modifications to headers might be too specific for this to be useful.

Also considered wrapping the framestacking routine as a function, but
runtime seemed to increase when I did so. Probably because it involves
running two list comprehensions, instead of one. I'm just going to leave
it here.

Any speed increase to the sigma clipping routine would be welcome. But
this issue has been known for a while in the python/scipi community, and
 none of them have found a solution yet that preserves the shape of the
data.