Wrote a wrapper function for the frame subtraction.
This function handles any mix of 3 use cases:
the frames to be subtracted are on disk, and passed as filenames
the frames are passed as HDULists, or simular objects
the frames are in DS9

Since DS9 only passes the HDU in the curretly loaded frame, this gives
rise to a special use case when one parameter is a ds9 instance, and the
 other is a file one disk or loaded in python.

A method was written to handle this case. This method takes an extension
 and finds a match in an hdu list.

Frame_subtract should be able to handle a list of HDUs instead of a
HDUList, but this has not been tested.

Frame_average was updated to include date average was done in the
primary HDU comments.