Changed _frame_mean to return an array that keeps track of data points used, simular to sigma_clipped average. Since all pixels are used in _frame_mean, this simply returns arrays with values equal to the number of frames used.
Cleaned up the code in _sigma_clipped_average. The runtime is the same,
but the code is easier to read and more pythonic.

Also wrote some more on the frame_average function docstring.

There is something wrong with how this averaging is running. Something
to do with how astropy.fits is opening the files. I tried turning off
memory mapping, this did not help, if anything it got worst.