Fixed the memory leak.
Acutally turned out to be two memory leaks. One was the issue with
pyfits. Their fix of using gc.collect() solved the issue. The other was
with how I was opening and closing files (I believe).

I changed how file opening and data unpacking was handled, both to fix
the memory leak, and because this way made more sense. Data unpacking
now happens in the wrapper function, and the data calculations are done
by calling a method.

Made various improvements to the code. This resulted in a slight
improvement in runtime.