Added error calculations to the naive average, sigma clipped average, and the median methods. It should be noted that the median errors are based upon the assuption of a normally distrubuted data set. Since medians are only use when the data is NOT normal, this assumption is false right out the gate, meaning the errors should not be trusted.
Updated the docstrings for the above functions. Moved garbage collection
to inside the functions. This garbage collection is nessecary to catch
any memory that was not properly deallocated by opening and closing fits
files quickly. However, the garbage collection may not have any effect
until the function has finished returning. This is problematic,
because the user should not have to worry about memory cleanup.

Updated get_filenames to default to the current directory if no path is
given, and updated the docstring to reflect this