Fixed the issues with the write to file routine, by having all functions pass the file path seperately. Updated the documentation to reflect this, and to include the write to file feature.
Cleaned up the code, especially in sigma_clipped_average. The method for
sigma clipping is much more pythonic now, using zip instead of
enumerate. Also caught some errors in how arguments were being passed.

Also commented out the gc.collect(), cause it doesn't seem to do
anything unless it's called after the script runs.

The functions seem to be working. But the memory usage is still
suspicious: the amount of memory being held open by pycharm increases
after the program finishes executing. But, this should clear when the
application closes, is less than 1Gb, and could be due to pycharm, not
the program, so I am ignoring it for now.